Added modules and restructured files
