+++ index.html
@@ -6,7 +6,7 @@
 		<meta http-equiv="X-UA-Compatible" content="ie=edge" />
 		<title>ProjectManager</title>
 		<link rel="stylesheet" href="app.css" />
-		<script src="dist/app.js" defer></script>
+		<script type="module" src="dist/app.js"></script>
 	</head>
 	<body>
 		<template id="project-input">
