+++ src/app.ts
@@ -1,379 +1,6 @@
-interface Draggable {
-	dragStartHandler(event: DragEvent): void;
-	dragEndHandler(event: DragEvent): void;
-}
+import { ProjectInput } from './components/project-input.js';
+import { ProjectList } from './components/project-list.js';
 
-interface DragTarget {
-	dragOverHandler(event: DragEvent): void;
-	dropHandler(event: DragEvent): void;
-	dragLeaveHandler(event: DragEvent): void;
-}
-
-enum ProjectStatus {
-	Active,
-	Finished,
-}
-class Project {
-	constructor(
-		public id: string,
-		public title: string,
-		public description: string,
-		public people: number,
-		public status: ProjectStatus
-	) {}
-}
-
-type Listener<T> = (items: T[]) => void;
-
-class State<T> {
-	protected listeners: Listener<T>[] = [];
-
-	addListener(listenerFn: Listener<T>) {
-		this.listeners.push(listenerFn);
-	}
-}
-
-class ProjectState extends State<Project> {
-	private projects: Project[] = [];
-	private static instance: ProjectState;
-
-	private constructor() {
-		super();
-	}
-
-	static getInstance() {
-		if (!this.instance) {
-			this.instance = new ProjectState();
-		}
-		return this.instance;
-	}
-
-	addProject(title: string, description: string, numOfPeole: number) {
-		const newProject = new Project(
-			Math.random().toString(),
-			title,
-			description,
-			numOfPeole,
-			ProjectStatus.Active
-		);
-		this.projects.push(newProject);
-		this.updateListeners();
-	}
-
-	moveProject(projectId: string, newStatus: ProjectStatus) {
-		const project = this.projects.find((prj) => prj.id === projectId);
-		if (project && project.status !== newStatus) {
-			project.status = newStatus;
-			this.updateListeners();
-		}
-	}
-
-	private updateListeners() {
-		for (const listenerFn of this.listeners) {
-			listenerFn(this.projects.slice());
-		}
-	}
-}
-
-const projectState = ProjectState.getInstance();
-
-interface Validatable {
-	value: string | number;
-	required: boolean;
-	maxLength?: number;
-	minLength?: number;
-	max?: number;
-	min?: number;
-}
-
-function validate(validatableInput: Validatable) {
-	let isValid = true;
-	if (validatableInput.required) {
-		isValid =
-			isValid && validatableInput.value.toString().trim().length !== 0;
-	}
-	if (
-		validatableInput.minLength != null &&
-		typeof validatableInput.value === 'string'
-	) {
-		isValid =
-			isValid &&
-			validatableInput.value.length > validatableInput.minLength;
-	}
-	if (
-		validatableInput.maxLength != null &&
-		typeof validatableInput.value === 'string'
-	) {
-		isValid =
-			isValid &&
-			validatableInput.value.length < validatableInput.maxLength;
-	}
-	if (
-		validatableInput.max != null &&
-		typeof validatableInput.value === 'number'
-	) {
-		isValid = isValid && validatableInput.value < validatableInput.max;
-	}
-	if (
-		validatableInput.min != null &&
-		typeof validatableInput.value === 'number'
-	) {
-		isValid = isValid && validatableInput.value > validatableInput.min;
-	}
-	return isValid;
-}
-
-function autoBind(_: any, _2: string, descriptor: PropertyDescriptor) {
-	const originalMethod = descriptor.value;
-	const adjDescriptor: PropertyDescriptor = {
-		configurable: true,
-		get() {
-			return originalMethod.bind(this);
-		},
-	};
-	return adjDescriptor;
-}
-
-//Component Base Class
-
-abstract class Component<T extends HTMLElement, U extends HTMLElement> {
-	templateElement: HTMLTemplateElement;
-	hostElement: T;
-	element: U;
-
-	constructor(
-		templateElemId: string,
-		hostElemId: string,
-		insertAtStart: boolean,
-		newElemId?: string
-	) {
-		this.templateElement = document.getElementById(
-			templateElemId
-		)! as HTMLTemplateElement;
-		this.hostElement = document.getElementById(hostElemId)! as T;
-
-		const importedNode = document.importNode(
-			this.templateElement.content,
-			true
-		);
-		this.element = importedNode.firstElementChild as U;
-		if (newElemId) {
-			this.element.id = newElemId;
-		}
-		this.attach(insertAtStart);
-	}
-
-	private attach(insertAtBeginning: boolean) {
-		this.hostElement.insertAdjacentElement(
-			insertAtBeginning ? 'afterbegin' : 'beforeend',
-			this.element
-		);
-	}
-
-	abstract configure(): void;
-	abstract renderContent(): void;
-}
-
-class ProjectItem extends Component<HTMLUListElement, HTMLLIElement>
-	implements Draggable {
-	private project: Project;
-
-	get persons() {
-		if (this.project.people === 1) {
-			return '1 person';
-		} else {
-			return `${this.project.people} persons`;
-		}
-	}
-
-	constructor(hostId: string, project: Project) {
-		super('single-project', hostId, false, project.id);
-		this.project = project;
-
-		this.configure();
-		this.renderContent();
-	}
-
-	@autoBind
-	dragStartHandler(event: DragEvent) {
-		event.dataTransfer!.setData('text/plain', this.project.id);
-		event.dataTransfer!.effectAllowed = 'move';
-	}
-
-	dragEndHandler(_: DragEvent) {
-		console.log('DragEnd');
-	}
-
-	configure() {
-		this.element.addEventListener('dragstart', this.dragStartHandler);
-		this.element.addEventListener('dragend', this.dragEndHandler);
-	}
-
-	renderContent() {
-		this.element.querySelector('h2')!.textContent = this.project.title;
-		this.element.querySelector(
-			'h3'
-		)!.textContent = `${this.persons} assigned`;
-		this.element.querySelector('p')!.textContent = this.project.description;
-	}
-}
-class ProjectList extends Component<HTMLDivElement, HTMLElement>
-	implements DragTarget {
-	assignedProjects: Project[];
-
-	constructor(private type: 'active' | 'finished') {
-		super('project-list', 'app', false, `${type}-projects`);
-		this.assignedProjects = [];
-
-		this.configure();
-		this.renderContent();
-	}
-
-	@autoBind
-	dragOverHandler(event: DragEvent) {
-		if (
-			event.dataTransfer &&
-			event.dataTransfer.types[0] === 'text/plain'
-		) {
-			event.preventDefault();
-			const listElem = this.element.querySelector('ul')!;
-			listElem.classList.add('droppable');
-		}
-	}
-
-	@autoBind
-	dropHandler(event: DragEvent) {
-		const prjId = event.dataTransfer!.getData('text/plain');
-		projectState.moveProject(
-			prjId,
-			this.type === 'active'
-				? ProjectStatus.Active
-				: ProjectStatus.Finished
-		);
-	}
-
-	@autoBind
-	dragLeaveHandler(_: DragEvent) {
-		const listElem = this.element.querySelector('ul')!;
-		listElem.classList.remove('droppable');
-	}
-
-	configure() {
-		this.element.addEventListener('dragover', this.dragOverHandler);
-		this.element.addEventListener('dragleave', this.dragLeaveHandler);
-		this.element.addEventListener('drop', this.dropHandler);
-
-		projectState.addListener((projects: Project[]) => {
-			const relevantProjects = projects.filter((prj) => {
-				if (this.type === 'active') {
-					return prj.status === ProjectStatus.Active;
-				}
-				return prj.status === ProjectStatus.Finished;
-			});
-
-			this.assignedProjects = relevantProjects;
-			this.renderProjects();
-		});
-	}
-
-	renderContent() {
-		const listId = `${this.type}-projects-list`;
-		this.element.querySelector('ul')!.id = listId;
-		this.element.querySelector('h2')!.textContent =
-			this.type.toUpperCase() + 'PROJECTS';
-	}
-
-	private renderProjects() {
-		const listElem = document.getElementById(
-			`${this.type}-projects-list`
-		)! as HTMLUListElement;
-		listElem.innerHTML = '';
-		for (const prjItem of this.assignedProjects) {
-			new ProjectItem(this.element.querySelector('ul')!.id, prjItem);
-		}
-	}
-}
-
-class ProjectInput extends Component<HTMLDivElement, HTMLFormElement> {
-	titleInputElement: HTMLInputElement;
-	descriptionInputElement: HTMLInputElement;
-	peopleInputElement: HTMLInputElement;
-
-	constructor() {
-		super('project-input', 'app', true, 'user-input');
-		this.titleInputElement = this.element.querySelector(
-			'#title'
-		) as HTMLInputElement;
-		this.descriptionInputElement = this.element.querySelector(
-			'#description'
-		) as HTMLInputElement;
-		this.peopleInputElement = this.element.querySelector(
-			'#people'
-		) as HTMLInputElement;
-
-		this.configure();
-	}
-
-	configure() {
-		this.element.addEventListener('submit', this.handleSubmit);
-	}
-
-	private gatherUserInput(): [string, string, number] | void {
-		const enteredTitle = this.titleInputElement.value;
-		const enteredDescription = this.descriptionInputElement.value;
-		const enteredPeople = this.peopleInputElement.value;
-
-		const titleValidateble = {
-			value: enteredTitle,
-			required: true,
-		};
-
-		const descriptionValidateble = {
-			value: enteredDescription,
-			required: true,
-			minLength: 5,
-		};
-
-		const peopleValidateble = {
-			value: enteredPeople,
-			required: true,
-			min: 1,
-			max: 5,
-		};
-
-		if (
-			!validate(titleValidateble) &&
-			!validate(descriptionValidateble) &&
-			!validate(peopleValidateble)
-		) {
-			console.log('Invalid Input, please try again!');
-			return;
-		} else {
-			return [enteredTitle, enteredDescription, +enteredPeople];
-		}
-	}
-
-	private clearInputs() {
-		this.titleInputElement.value = '';
-		this.descriptionInputElement.value = '';
-		this.peopleInputElement.value = '';
-	}
-
-	@autoBind
-	private handleSubmit(event: Event) {
-		event.preventDefault();
-		const userInput = this.gatherUserInput();
-		if (Array.isArray(userInput)) {
-			const [title, desc, people] = userInput;
-			projectState.addProject(title, desc, people);
-			console.log(title, desc, people);
-			this.clearInputs();
-		}
-	}
-
-	renderContent() {}
-}
-
-const projImput = new ProjectInput();
-const activeProjectList = new ProjectList('active');
-const finishedProjectList = new ProjectList('finished');
+new ProjectInput();
+new ProjectList('active');
+new ProjectList('finished');
