+++ src/decorators/autobind.ts
@@ -1,0 +1,10 @@
+export function autoBind(_: any, _2: string, descriptor: PropertyDescriptor) {
+	const originalMethod = descriptor.value;
+	const adjDescriptor: PropertyDescriptor = {
+		configurable: true,
+		get() {
+			return originalMethod.bind(this);
+		},
+	};
+	return adjDescriptor;
+}
