+++ src/state/project-state.ts
@@ -1,0 +1,55 @@
+import { Project, ProjectStatus } from '../models/project.js';
+
+type Listener<T> = (items: T[]) => void;
+
+class State<T> {
+	protected listeners: Listener<T>[] = [];
+
+	addListener(listenerFn: Listener<T>) {
+		this.listeners.push(listenerFn);
+	}
+}
+
+export class ProjectState extends State<Project> {
+	private projects: Project[] = [];
+	private static instance: ProjectState;
+
+	private constructor() {
+		super();
+	}
+
+	static getInstance() {
+		if (!this.instance) {
+			this.instance = new ProjectState();
+		}
+		return this.instance;
+	}
+
+	addProject(title: string, description: string, numOfPeole: number) {
+		const newProject = new Project(
+			Math.random().toString(),
+			title,
+			description,
+			numOfPeole,
+			ProjectStatus.Active
+		);
+		this.projects.push(newProject);
+		this.updateListeners();
+	}
+
+	moveProject(projectId: string, newStatus: ProjectStatus) {
+		const project = this.projects.find((prj) => prj.id === projectId);
+		if (project && project.status !== newStatus) {
+			project.status = newStatus;
+			this.updateListeners();
+		}
+	}
+
+	private updateListeners() {
+		for (const listenerFn of this.listeners) {
+			listenerFn(this.projects.slice());
+		}
+	}
+}
+
+export const projectState = ProjectState.getInstance();
